+++ packages/leancode_lint/lib/helpers.dart
@@ -28,3 +28,47 @@ Expression? maybeGetSingleReturnExpression(FunctionBody body) {
     _ => null,
   };
 }
+
+/// Returns all return expressions from passed statement recursively.
+Iterable<Expression> getAllInnerReturnStatements(Statement statement) {
+  switch (statement) {
+    case IfStatement():
+      return [
+        ...getAllInnerReturnStatements(statement.thenStatement),
+        if (statement.elseStatement case final statement?)
+          ...getAllInnerReturnStatements(statement),
+      ];
+
+    case ForStatement(:final body):
+      return getAllInnerReturnStatements(body);
+
+    case Block():
+      return statement.statements.expand(getAllInnerReturnStatements);
+
+    case TryStatement(:final body, :final catchClauses, :final finallyBlock):
+      return [
+        ...body.statements.expand(getAllInnerReturnStatements),
+        ...catchClauses
+            .expand((clause) => clause.body.statements)
+            .expand(getAllInnerReturnStatements),
+        ...?finallyBlock?.statements.expand(getAllInnerReturnStatements),
+      ];
+
+    case DoStatement(:final body):
+      return getAllInnerReturnStatements(body);
+
+    case WhileStatement(:final body):
+      return getAllInnerReturnStatements(body);
+
+    case ExpressionStatement():
+      return getAllInnerReturnStatements(statement);
+
+    case ReturnStatement():
+      return [
+        if (statement.expression case final expression?) expression,
+      ];
+
+    default:
+      return [];
+  }
+}