+++ packages/leancode_lint/lib/leancode_lint.dart
@@ -1,9 +1,8 @@
-// This is the entrypoint of our custom linter
-
 import 'package:custom_lint_builder/custom_lint_builder.dart';
-import 'package:leancode_lint/custom_lints/add_sliver_prefix_for_widget_returning_sliver.dart';
-import 'package:leancode_lint/custom_lints/start_comment_with_space.dart';
-import 'package:leancode_lint/custom_lints/use_design_system_item.dart';
+import 'package:leancode_lint/assists/convert_record_into_nominal_type.dart';
+import 'package:leancode_lint/lints/add_sliver_prefix_for_widget_returning_sliver.dart';
+import 'package:leancode_lint/lints/start_comment_with_space.dart';
+import 'package:leancode_lint/lints/use_design_system_item.dart';
 
 PluginBase createPlugin() => _Linter();
 
@@ -11,10 +10,12 @@ class _Linter extends PluginBase {
   @override
   List<LintRule> getLintRules(CustomLintConfigs configs) => [
         StartCommentWithSpace(),
-        ...?UseDesignSystemItem.getRulesListFromConfigs(configs),
+        ...UseDesignSystemItem.getRulesListFromConfigs(configs),
         AddSliverPrefixForWidgetReturningSliver(),
       ];
 
   @override
-  List<Assist> getAssists() => [];
+  List<Assist> getAssists() => [
+        ConvertRecordIntoNominalType(),
+      ];
 }