+++ packages/leancode_lint/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 7.0.0
+
+- Implement LeanCode custom lints
+
 # 6.0.0
 
 - Enable the following lints: